requirements.txt
# for databricks
pyspark

# for Azure Event Hub
azure-eventhub

# for CosmosDB
azure-cosmos

# for cli
azure-cli>=2.12.1